+++ product-shopthethao/ShopeTheThao_frontend_react/src/components/Admin/Products/FormFields/ImageUpload.jsx
@@ -1,26 +1,10 @@
-// import React from 'react';
-// import { Form, Upload, Row } from 'antd';
-// import { productValidationRules } from 'utils/formValidation';
+import React from 'react';
+import { Form, Upload, Row } from 'antd';
+import { productValidationRules } from 'utils/formValidation';
 
-// const ImageUpload = ({ FileList, handleUploadChange }) => (
-//   <Row gutter={16} justify="space-between">
-//     <Form.Item
-//       label="Hình ảnh sản phẩm"
-//       name="images"
-//       rules={productValidationRules.images}
-//     >
-//       <Upload
-//         beforeUpload={() => false}
-//         accept=".png, .jpg, .jpeg"
-//         listType="picture-card"
-//         fileList={FileList}
-//         onChange={handleUploadChange}
-//         multiple
-//       >
-//         {FileList.length < 5 && "+ Upload"}
-//       </Upload>
-//     </Form.Item>
-//   </Row>
-// );
+const ImageUpload = ({ FileList, handleUploadChange }) => (
+<>
+</>
+);
 
-// export default ImageUpload;
+export default ImageUpload;
