+++ product-shopthethao/ShopeTheThao_frontend_react/src/pages/Admin/Products/index.jsx
@@ -1,25 +1,16 @@
 import React, { useState, useEffect, useRef } from "react";
 import {
   Button,
-  Input,
   Space,
   Form,
   Modal,
   message,
-  Image,
-  Tag,
-  Tooltip,
   Select,
   Table,
   Row,
-  Upload,
-  Col,
 } from "antd";
 import {
-  MinusCircleOutlined,
   PlusOutlined,
-  CheckCircleOutlined,
-  CloseCircleOutlined,
   EditOutlined,
 } from "@ant-design/icons";
 
@@ -27,9 +18,10 @@ import uploadApi from "api/service/uploadApi";
 import PaginationComponent from "components/PaginationComponent";
 import { useCategories, useSizes } from "hooks";
 import { productsApi } from "api/Admin";
-import "..//index.scss";
-import styles from "..//modalStyles.module.scss";
-import ActionColumn from "components/Admin/tableColumns/ActionColumn";
+import "../index.scss";
+import styles from "../modalStyles.module.scss";
+import { ProductColumns, ProductForm } from "components/Admin";
+
 
 const Products = () => {
   const [open, setOpen] = useState(false);
@@ -264,7 +256,7 @@ const Products = () => {
     try {
       await productsApi.delete(id);
       message.success("Xóa sản phẩm thành công!");
-      setWorkSomeThing([!workSomeThing]);
+      setWorkSomeThing(!workSomeThing);
       setProducts(products.filter((p) => p.id !== id));
     } catch (error) {
       message.error("Không thể xóa sản phẩm!");
@@ -375,136 +367,8 @@ const Products = () => {
     setCurrentPage(1);
   };
 
-  // Cấu hình cột bảng
-  const columns = [
-    { title: "🆔 ID", dataIndex: "id", key: "id" },
-    {
-      title: "🏷️ Tên sản phẩm",
-      dataIndex: "name",
-      key: "name",
-      render: (text) => (
-        <Tooltip title={text || "Không có mô tả"} placement="top">
-          <span className="ellipsis-text">
-            {text?.length > 35
-              ? `${text.substring(0, 15)}...`
-              : text || "Không có Tên sản phẩm"}
-          </span>
-        </Tooltip>
-      ),
-    },
-    {
-      title: "📦 Số Lượng",
-      dataIndex: "totalQuantity",
-      key: "totalQuantity",
-    },
-    {
-      title: "📂 Loại sản phẩm",
-      dataIndex: ["categorie", "name"],
-      key: "categorie",
-      render: (text) => (
-        <Tooltip title={text || "Không có mô tả"} placement="top">
-          <span className="ellipsis-text">
-            {text?.length > 25
-              ? `${text.substring(0, 15)}...`
-              : text || "Không có mô tả"}
-          </span>
-        </Tooltip>
-      ),
-    },
-    {
-      title: "📝 Mô tả sản phẩm",
-      dataIndex: "description",
-      key: "description",
-      render: (text) => (
-        <Tooltip title={text || "Không có mô tả"} placement="top">
-          <span className="ellipsis-text">
-            {text?.length > 20
-              ? `${text.substring(0, 20)}...`
-              : text || "Không có mô tả"}
-          </span>
-        </Tooltip>
-      ),
-    },
-    {
-      title: "📊 Trạng thái",
-      dataIndex: "totalQuantity",
-      key: "status",
-      render: (totalQuantity) => (
-        <Tag
-          icon={
-            totalQuantity > 0 ? (
-              <CheckCircleOutlined />
-            ) : (
-              <CloseCircleOutlined />
-            )
-          }
-          color={totalQuantity > 0 ? "green" : "red"}
-          style={{
-            borderRadius: "12px",
-            padding: "4px 12px",
-            transition: "all 0.3s ease",
-            cursor: "pointer",
-          }}
-          onMouseEnter={(e) => {
-            e.target.style.transform = "scale(1.1)";
-          }}
-          onMouseLeave={(e) => {
-            e.target.style.transform = "scale(1)";
-          }}
-        >
-          {totalQuantity > 0 ? "Còn sản phẩm" : "Hết sản phẩm"}
-        </Tag>
-      ),
-    },
-
-    {
-      title: "🖼️ Ảnh sản phẩm",
-      dataIndex: "images",
-      key: "images",
-      render: (_, record) => (
-        <Space size="middle">
-          {record.images && record.images.length > 0 ? (
-            record.images.map((image, index) => (
-              <Image
-                key={index}
-                width={80}
-                height={80}
-                style={{ objectFit: "contain" }}
-                src={`http://localhost:8081/api/upload/${image.imageUrl}`}
-                alt="Ảnh sản phẩm"
-              />
-            ))
-          ) : (
-            <span>Không có ảnh</span>
-          )}
-        </Space>
-      ),
-    },
-    {
-      title: "💵 Giá Mặc định",
-      dataIndex: "price",
-      key: "price",
-      render: (price) => `${price.toLocaleString()} VND`,
-    },
-    {
-      title: "Kích cỡ | Số Lượng | Giá tiền",
-
-      dataIndex: "sizes",
-      key: "sizes",
-      render: (sizes) => (
-        <Space direction="vertical" size="small">
-          {sizes.map((size, index) => (
-            <div key={index}>
-              <strong>{size.size?.name}</strong> - {size.quantity} Sản Phẩm -{" "}
-              {size.price.toLocaleString()} VND
-            </div>
-          ))}
-        </Space>
-      ),
-    },
-
-    ActionColumn(handleEditData, handleDelete),
-  ];
+  // Sử dụng component ProductColumns
+  const columns = ProductColumns(handleEditData, handleDelete);
 
   return (
     <div style={{ padding: 10 }}>
@@ -567,236 +431,17 @@ const Products = () => {
           }}
           cancelText="Hủy"
         >
-          <Form 
+          <ProductForm 
             form={form} 
-            layout="vertical"
-            onValuesChange={handleFormValuesChange}
-          >
-            <Row gutter={16}>
-              <Col span={24}>
-                <Form.Item
-                  name="name"
-                  label="Tên sản phẩm"
-                  rules={[
-                    { required: true, message: "Vui lòng nhập tên sản phẩm!" },
-                  ]}
-                >
-                  <Input placeholder="Nhập tên sản phẩm" />
-                </Form.Item>
-              </Col>
-            </Row>
-
-            <Row gutter={16}>
-              <Col span={24}>
-                <Form.Item
-                  name="description"
-                  label="Mô tả sản phẩm"
-                  rules={[
-                    {
-                      required: true,
-                      message: "Vui lòng nhập mô tả sản phẩm!",
-                    },
-                  ]}
-                >
-                  <Input.TextArea rows={2} placeholder="Nhập mô tả sản phẩm" />
-                </Form.Item>
-              </Col>
-            </Row>
-
-            <Row gutter={16}>
-              <Col span={12}>
-                <Form.Item name="totalQuantity" label="Tổng số lượng">
-                  <Input value={totalQuantity} disabled />
-                </Form.Item>
-              </Col>
-
-              <Col span={12}>
-                <Form.Item
-                  name="categorie"
-                  label="Chọn danh mục"
-                  rules={[
-                    { required: true, message: "Vui lòng chọn danh mục" },
-                  ]}
-                >
-                  <Select
-                    style={{ width: "100%" }}
-                    showSearch
-                    placeholder="Chọn danh mục"
-                    optionFilterProp="children"
-                    filterOption={(input, option) =>
-                      (option?.label ?? "")
-                        .toLowerCase()
-                        .includes(input.toLowerCase())
-                    }
-                    options={categories?.map((categorie) => ({
-                      value: categorie.id,
-                      label: categorie.name,
-                    }))}
-                  />
-                </Form.Item>
-              </Col>
-            </Row>
-
-            {/* Upload ảnh */}
-            <Row gutter={16} justify="space-between">
-              <Form.Item
-                label="Hình ảnh sản phẩm"
-                name="images"
-                rules={[
-                  {
-                    required: true,
-                    message: "Vui lòng tải lên ít nhất một hình ảnh!",
-                  },
-                ]}
-              >
-                <Upload
-                  beforeUpload={() => false}
-                  accept=".png, .jpg, .jpeg"
-                  listType="picture-card"
-                  fileList={FileList}
-                  onChange={handleUploadChange}
-                  multiple
-                >
-                  {FileList.length < 5 && "+ Upload"}
-                </Upload>
-              </Form.Item>
-            </Row>
-
-            <Row gutter={16}>
-              <Col span={24}>
-                <Form.Item
-                  name="price"
-                  label="Giá sản phẩm"
-                  rules={[
-                    { required: true, message: "Vui lòng nhập giá sản phẩm!" },
-                    {
-                      validator: (_, value) => {
-                        if (!value || isNaN(value) || value < 1000) {
-                          return Promise.reject(
-                            new Error(
-                              "Giá sản phẩm không thể nhỏ hơn 1,000 VND!"
-                            )
-                          );
-                        }
-                        if (value > 1000000000) {
-                          return Promise.reject(
-                            new Error(
-                              "Giá sản phẩm không thể vượt quá 1 tỷ VND!"
-                            )
-                          );
-                        }
-                        return Promise.resolve();
-                      },
-                    },
-                  ]}
-                >
-                  <Input
-                    type="number"
-                    min={1000}
-                    max={1000000000}
-                    step={1000}
-                    placeholder="Nhập giá sản phẩm (VND)"
-                  />
-                </Form.Item>
-              </Col>
-            </Row>
-            <Form.List
-              name="sizes"
-              initialValue={
-                editingProduct
-                  ? editingProduct.sizes.map((size) => ({
-                      size: size.size.id,
-                      quantity: size.quantity,
-                      price: size.price,
-                    }))
-                  : []
-              }
-            >
-              {(fields, { add, remove }) => (
-                <>
-                  {fields.map(({ key, name, fieldKey, ...restField }) => (
-                    <Row key={key} gutter={16}>
-                      <Col span={8}>
-                        <Form.Item
-                          {...restField}
-                          name={[name, "size"]}
-                          label="Kích cỡ"
-                          rules={[
-                            {
-                              required: true,
-                              message: "Vui lòng chọn kích cỡ!",
-                            },
-                          ]}
-                        >
-                          <Select
-                            options={sizes.map((size) => ({
-                              value: size.id,
-                              label: size.name,
-                            }))}
-                            onChange={(value) => handleSizeChange(value, name)} // Đảm bảo mỗi lần thay đổi gọi hàm kiểm tra trùng
-                          />
-                        </Form.Item>
-                      </Col>
-
-                      <Col span={8}>
-                        <Form.Item
-                          {...restField}
-                          name={[name, "quantity"]}
-                          label="Số lượng"
-                          rules={[
-                            {
-                              required: true,
-                              message: "Vui lòng nhập số lượng lớn hơn 1!",
-                            },
-                          ]}
-                        >
-                          <Input
-                            type="number"
-                            min={1}
-                            onChange={(e) => {
-                              const value = Math.max(0, e.target.value);
-                              handleSizeQuantityChange(value, name);
-                            }}
-                          />
-                        </Form.Item>
-                      </Col>
-
-                      <Col span={8}>
-                        <Form.Item
-                          {...restField}
-                          name={[name, "price"]}
-                          label="Giá"
-                        >
-                          <Input type="number" min={0} />
-                        </Form.Item>
-                      </Col>
-
-                      <Col span={24}>
-                        <Button
-                          type="danger"
-                          onClick={() => remove(name)}
-                          icon={<MinusCircleOutlined />}
-                          block
-                        >
-                          Xoá kích cỡ
-                        </Button>
-                      </Col>
-                    </Row>
-                  ))}
-                  <Form.Item>
-                    <Button
-                      type="dashed"
-                      onClick={() => add()}
-                      icon={<PlusOutlined />}
-                      block
-                    >
-                      Thêm kích cỡ
-                    </Button>
-                  </Form.Item>
-                </>
-              )}
-            </Form.List>
-          </Form>
+            categories={categories}
+            sizes={sizes}
+            totalQuantity={totalQuantity}
+            FileList={FileList}
+            handleUploadChange={handleUploadChange}
+            handleSizeChange={handleSizeChange}
+            handleSizeQuantityChange={handleSizeQuantityChange}
+            handleFormValuesChange={handleFormValuesChange}
+          />
           <Modal
             open={previewOpen}
             footer={null}
